Include the code in the singularity container.

+++ utils/make_submission.py
@@ -7,38 +7,46 @@ from zipfile import ZipFile
 from os.path import basename
 import shutil
 import pathlib
+from dotenv import load_dotenv
+import os
+load_dotenv()
+root = os.environ.get("root_folder")
 
+submission_type = '.sif'
 load_dotenv()
 
 submission_name = os.environ.get("submission_name")
 
 # Save everything in a submission_name folder in order to keep track of submissions
-outdir = "./output/" + submission_name + "/"
+outdir = root + "/output/" + submission_name + "/"
 p = pathlib.Path(outdir)
 p.mkdir(parents=True, exist_ok=True)
-shutil.copyfile("./src/model.py", outdir + "model.py")
-shutil.copyfile("./singularity/container.sif", outdir + submission_name + ".sif")
+shutil.copyfile(root +"/src/model.py", outdir + "model.py")
+shutil.copyfile(root +"/singularity/container.sif", outdir + submission_name + ".sif")
 
 
-outdir = "./output/" + submission_name + "/output/"
+outdir = root + "/output/" + submission_name + "/output/"
 p = pathlib.Path(outdir)
 p.mkdir(parents=True, exist_ok=True)
-shutil.copyfile("./output/score.csv", outdir + "score.csv")
+shutil.copyfile(root + "/output/score.csv", outdir + "score.csv")
 
-file_to_upload = "./output/" + submission_name + '/' + submission_name + '.zip' 
-try:
-    os.remove(file_to_upload)
-except OSError:
-    pass
+if submission_type == ".zip":
+    file_to_upload = root + "/output/" + submission_name + '/' + submission_name + '.zip' 
+    try:
+        os.remove(file_to_upload)
+    except OSError:
+        pass
 
-to_zip = "./output/" + submission_name +'/'
-zip_file =  "./output/" + submission_name
-shutil.make_archive(zip_file, "zip",to_zip)
+    to_zip = root + "/output/" + submission_name +'/'
+    zip_file =  root + "/output/" + submission_name
+    shutil.make_archive(zip_file, "zip",to_zip)
 
-shutil.move(
-    zip_file + '.zip', file_to_upload
-)  # In two steps in order to avoid putting a zip in a zip
-
+    shutil.move(
+        zip_file + '.zip', file_to_upload
+    )  # In two steps in order to avoid putting a zip in a zip
+elif submission_type == ".sif":
+    file_to_upload = root + "/output/" + submission_name + '/' + submission_name + '.sif'
+      
 username = os.environ.get("username")
 print(username)
 password = os.environ.get("password")
@@ -49,8 +57,9 @@ syn = synapseclient.login(username, password)
 #%%
 # Add a local file to an existing project on Synapse
 file = File(path=file_to_upload, parent=project_id)
-#file = syn.store(file) ####################################################################################################""""
+file = syn.store(file) ####################################################################################################""""
 
 # syn.delete(file) # API documentation : https://help.synapse.org/docs/API-Clients-and-Documentation.1985446128.html
 
  
+# %%
